requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
fonttools==4.37.4
future==0.18.2
greenlet==1.1.3
idna==3.4
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.0
numpy==1.23.3
obspy==1.3.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scipy==1.9.1
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
urllib3==1.26.12
